+++ pkg/consts/agent_injection.go
@@ -8,7 +8,8 @@ const (
 	AgentInitSecretName                = "dynatrace-dynakube-config"
 	AgentInitSecretConfigField         = "config"
 
-	BootstrapperInitSecretName = "dynatrace-bootstrapper-config"
+	BootstrapperInitSecretName      = "dynatrace-bootstrapper-config"
+	BootstrapperInitCertsSecretName = "dynatrace-bootstrapper-certs
 
 	LdPreloadFilename = "ld.so.preload"
 	LibAgentProcPath  = "/agent/lib64/liboneagentproc.so"
